Created and began working on a wolfSSL-based acvp_app
Unfortunately, it doesn't work quite yet.  The whole team worked on
this (on a large screen), and we can't figure out why the server is
failing us.  Currently, we only support SHA256.  From our testing, we
are seeing many instances of the same vector, giving the same digest,
but every test fails.