Began updating curl_easy_perform to WolfSSL
Removed OpenSSL functionality from the first half or so of
curl_easy_perform, and replaced it with WolfSSL. X509 certificate
purpose doesn't exist so we can't swap that over, it's been commented
out instead. All other X509 stuff should now be swapped to WolfSSL
equivalents. We made it as far as the "end wolfssl converted" comment on
line 773 and will pick up there later.